Only expose interface for MatrixClient
