+++ src/main/java/org/synyx/matrix/bot/MatrixClient.java
@@ -1,60 +1,28 @@
 package org.synyx.matrix.bot;
 
-import com.fasterxml.jackson.databind.DeserializationFeature;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.SerializationFeature;
-import com.fasterxml.jackson.databind.json.JsonMapper;
-import com.fasterxml.jackson.datatype.jdk8.Jdk8Module;
-import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-import lombok.extern.slf4j.Slf4j;
 import org.synyx.matrix.bot.domain.MatrixEventId;
 import org.synyx.matrix.bot.domain.MatrixRoomId;
-import org.synyx.matrix.bot.domain.MatrixUserId;
-import org.synyx.matrix.bot.internal.MatrixAuthentication;
-import org.synyx.matrix.bot.internal.MatrixBackoffException;
-import org.synyx.matrix.bot.internal.MatrixEventNotifier;
-import org.synyx.matrix.bot.internal.MatrixStateSynchronizer;
-import org.synyx.matrix.bot.internal.api.MatrixApi;
-import org.synyx.matrix.bot.internal.api.MatrixApiException;
-import org.synyx.matrix.bot.internal.api.dto.MessageDto;
-import org.synyx.matrix.bot.internal.api.dto.ReactionDto;
-import org.synyx.matrix.bot.internal.api.dto.ReactionRelatesToDto;
-import org.synyx.matrix.bot.internal.api.dto.SyncResponseDto;
+import org.synyx.matrix.bot.internal.MatrixClientImpl;
 
-import java.io.IOException;
 import java.util.Optional;
 
-@Slf4j
-public class MatrixClient {
+/**
+ * An interface for a client connecting to a matrix server.
+ * Serves as the main method of communicating with the server.
+ */
+public interface MatrixClient {
 
-  private static final long DEFAULT_BACKOFF_IN_SEC = 3;
-  private static final long BACKOFF_MAX_IN_SEC = 60;
+  /**
+   * Creates a new matrix client to connect to the specified server.
+   *
+   * @param url      The url for connecting to the intended matrix server. Must start with http:// or https://
+   * @param username The username for logging into the matrix server.
+   * @param password The password for logging into the matrix server.
+   * @return A {@link MatrixClient} implementation that connects to the specified matrix server.
+   */
+  static MatrixClient create(String url, String username, String password) {
 
-  private final MatrixAuthentication authentication;
-  private final ObjectMapper objectMapper;
-  private final MatrixApi api;
-  private MatrixState state;
-  private MatrixStateSynchronizer stateSynchronizer;
-  private MatrixPersistedStateProvider persistedState;
-  private MatrixEventNotifier eventNotifier;
-  private boolean interruptionRequested;
-  private long currentBackoffInSec;
-
-  public MatrixClient(String hostname, String username, String password) {
-
-    this.authentication = new MatrixAuthentication(username, password);
-    this.objectMapper = JsonMapper.builder()
-        .addModule(new Jdk8Module())
-        .addModule(new JavaTimeModule())
-        .configure(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS, false)
-        .configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false)
-        .configure(DeserializationFeature.READ_UNKNOWN_ENUM_VALUES_USING_DEFAULT_VALUE, true)
-        .build();
-    this.api = new MatrixApi(hostname, authentication, objectMapper);
-    this.state = null;
-    this.eventNotifier = null;
-    this.interruptionRequested = false;
-    this.currentBackoffInSec = DEFAULT_BACKOFF_IN_SEC;
+    return new MatrixClientImpl(url, username, password);
   }
 
   /**
@@ -64,10 +32,7 @@ public class MatrixClient {
    *
    * @param eventConsumer The consumer to call on events.
    */
-  public void setEventCallback(MatrixEventConsumer eventConsumer) {
-
-    this.eventNotifier = MatrixEventNotifier.from(objectMapper, eventConsumer).orElse(null);
-  }
+  void setEventCallback(MatrixEventConsumer eventConsumer);
 
   /**
    * Optionally provides an interface to provide the current state of the matrix client.
@@ -76,20 +41,8 @@ public class MatrixClient {
    *
    * @param persistedState An interface for persisting the matrix client state
    */
-  public void setPersistedStateProvider(MatrixPersistedStateProvider persistedState) {
+  void setPersistedStateProvider(MatrixPersistedStateProvider persistedState);
 
-    this.persistedState = persistedState;
-  }
-
-  /**
-   * Requests the matrix client to stop syncing and terminate.
-   * May be called from a different thread.
-   */
-  public void requestStopOfSync() {
-
-    interruptionRequested = true;
-    api.terminateOpenConnections();
-  }
 
   /**
    * The main matrix client event loop that continuously syncs all events happening on the matrix server to the client.
@@ -97,119 +50,28 @@ public class MatrixClient {
    *
    * @throws InterruptedException The sync has been interrupted
    */
-  public void syncContinuous() throws InterruptedException {
+  void syncContinuous() throws InterruptedException;
 
-    while (!interruptionRequested) {
-      try {
-        if (!authentication.isAuthenticated()) {
-          try {
-            api.login();
-          } catch (IOException e) {
-            throw new MatrixBackoffException("Failed to login to matrix server!", e);
-          } catch (MatrixApiException e) {
-            throw new MatrixCommunicationException("Failed to login to matrix server!", e);
-          }
+  /**
+   * Requests the matrix client to stop syncing and terminate.
+   * May be called from a different thread.
+   */
+  void requestStopOfSync();
 
-          log.info("Successfully logged in to matrix server as {}",
-              authentication.getUserId()
-                  .map(MatrixUserId::toString)
-                  .orElse("UNKNOWN")
-          );
-        }
-
-        state = new MatrixState(authentication.getUserId().orElseThrow(IllegalStateException::new));
-        stateSynchronizer = new MatrixStateSynchronizer(state, objectMapper);
-
-        SyncResponseDto syncResponse;
-        try {
-          syncResponse = api.syncFull()
-              .orElseThrow(() -> new MatrixCommunicationException("No data in initial sync"));
-        } catch (MatrixApiException | IOException e) {
-          throw new MatrixBackoffException("Failed to perform initial sync", e);
-        }
-
-        String lastBatch = syncResponse.nextBatch();
-        stateSynchronizer.synchronizeState(syncResponse);
-
-        if (eventNotifier != null) {
-          eventNotifier.getConsumer().onConnected(state);
-        }
-
-        if (persistedState != null) {
-          final var maybePersistedLastBatch = persistedState.getLastBatch();
-          if (maybePersistedLastBatch.isPresent()) {
-            lastBatch = maybePersistedLastBatch.get();
-          } else {
-            persistedState.setLastBatch(lastBatch);
-          }
-        }
-
-        while (!interruptionRequested) {
-          Optional<SyncResponseDto> maybePartialSyncResponse;
-
-          try {
-            maybePartialSyncResponse = api.sync(lastBatch);
-          } catch (MatrixApiException | IOException e) {
-            throw new MatrixBackoffException("Could not partial sync", e);
-          }
-
-          if (maybePartialSyncResponse.isPresent()) {
-            syncResponse = maybePartialSyncResponse.get();
-            lastBatch = syncResponse.nextBatch();
-
-            stateSynchronizer.synchronizeState(syncResponse);
-
-            if (eventNotifier != null) {
-              eventNotifier.notifyFromSynchronizationResponse(state, syncResponse);
-            }
-
-            if (persistedState != null) {
-              persistedState.setLastBatch(lastBatch);
-            }
-          }
-
-          currentBackoffInSec = DEFAULT_BACKOFF_IN_SEC;
-        }
-
-      } catch (MatrixBackoffException e) {
-        log.warn("Sync failed: {}, backing off for {}s", e.getCause().getClass().getName(), currentBackoffInSec);
-
-        clearSyncState();
-        Thread.sleep(currentBackoffInSec * 1000);
-        currentBackoffInSec = Math.min(currentBackoffInSec * 2, BACKOFF_MAX_IN_SEC);
-      }
-    }
-
-    clearSyncState();
-    interruptionRequested = false;
-    currentBackoffInSec = DEFAULT_BACKOFF_IN_SEC;
-  }
-
-  private void clearSyncState() {
-
-    authentication.clear();
-    state = null;
-  }
 
   /**
    * Returns whether the matrix client is currently connected to the server or not.
    *
    * @return {@code true} if the client is currently connected to the server, {@code false} otherwise.
    */
-  public boolean isConnected() {
-
-    return state != null;
-  }
+  boolean isConnected();
 
   /**
    * Returns the current state of the matrix client.
    *
    * @return A {@link MatrixState} object if currently connected to a server, {@link Optional#empty()} otherwise.
    */
-  public Optional<MatrixState> getState() {
-
-    return Optional.ofNullable(state);
-  }
+  Optional<MatrixState> getState();
 
   /**
    * Attempts to send a message to the specified room.
@@ -218,20 +80,7 @@ public class MatrixClient {
    * @param messageBody The body of the message to send.
    * @return A {@link MatrixEventId} containing the id of the event that was sent or {@link Optional#empty()} if sending the message did not succeed.
    */
-  public Optional<MatrixEventId> sendMessage(MatrixRoomId roomId, String messageBody) {
-
-    try {
-      return MatrixEventId.from(
-          api.sendEvent(roomId.getFormatted(), "m.room.message", new MessageDto(messageBody, "m.text"))
-      );
-    } catch (InterruptedException | IOException e) {
-      log.error("Failed to send message", e);
-    } catch (MatrixApiException e) {
-      log.warn("Could not send message", e);
-    }
-
-    return Optional.empty();
-  }
+  Optional<MatrixEventId> sendMessage(MatrixRoomId roomId, String messageBody);
 
   /**
    * Attempts to add a reaction to an event (a message of the time).
@@ -241,21 +90,7 @@ public class MatrixClient {
    * @param reaction The reaction to send.
    * @return A {@link MatrixEventId} containing the id of the event that was sent or {@link Optional#empty()} if sending the reaction did not succeed.
    */
-  public Optional<MatrixEventId> addReaction(MatrixRoomId roomId, MatrixEventId eventId, String reaction) {
-
-    final var reactionDto = new ReactionDto(new ReactionRelatesToDto(eventId.getFormatted(), reaction));
-    try {
-      return MatrixEventId.from(
-          api.sendEvent(roomId.getFormatted(), "m.reaction", reactionDto)
-      );
-    } catch (InterruptedException | IOException e) {
-      log.error("Failed to add reaction", e);
-    } catch (MatrixApiException e) {
-      log.warn("Could not add reaction", e);
-    }
-
-    return Optional.empty();
-  }
+  Optional<MatrixEventId> addReaction(MatrixRoomId roomId, MatrixEventId eventId, String reaction);
 
   /**
    * Attempts to join a room.
@@ -263,19 +98,7 @@ public class MatrixClient {
    * @param roomId The id of the room to join.
    * @return {@code true} if joining the room was successful, {@code false} otherwise.
    */
-  public boolean joinRoom(MatrixRoomId roomId) {
-
-    try {
-      api.joinRoom(roomId.getFormatted(), "i'm a bot");
-      return true;
-    } catch (InterruptedException | IOException e) {
-      log.error("Failed to join room", e);
-    } catch (MatrixApiException e) {
-      log.warn("Could not join room", e);
-    }
-
-    return false;
-  }
+  boolean joinRoom(MatrixRoomId roomId);
 
   /**
    * Attempts to leave a room.
@@ -283,17 +106,5 @@ public class MatrixClient {
    * @param roomId The id of the room to leave.
    * @return {@code true} if leaving the room was successful, {@code false} otherwise.
    */
-  public boolean leaveRoom(MatrixRoomId roomId) {
-
-    try {
-      api.leaveRoom(roomId.getFormatted(), "i'm a bot");
-      return true;
-    } catch (InterruptedException | IOException e) {
-      log.error("Failed to leave room", e);
-    } catch (MatrixApiException e) {
-      log.warn("Could not leave room", e);
-    }
-
-    return false;
-  }
+  boolean leaveRoom(MatrixRoomId roomId);
 }
